Req changes paolo

+++ src/components/ChangePassword/functions.js
@@ -1,7 +1,6 @@
-import swal from "sweetalert";
-import { API } from "../../services/env";
-import axios from "axios";
 import Validator from "../../helpers/Validations";
+import { put_request } from "../../helpers/Request";
+import swal from "sweetalert";
 
 export function validatePassword() {
   let error = false;
@@ -30,27 +29,28 @@ export function validatePassword() {
 export async function changePassword() {
   if (this.validatePassword()) {
     const passwords = {
-      token: localStorage.getItem("token"),
       oldPassword: this.state.old_password,
       newPassword: this.state.new_password,
     };
-    const res = await axios.put(`${API}/updatePassword`, passwords);
-    const msg = res.data.msg;
-    if (msg !== "Error") {
-      swal(
-        "¡Listo!",
-        "Contraseña actualizada exitosamente. Por motivos de seguridad, se cerrará su sesión",
-        "success"
-      ).then(() => {
-        localStorage.clear();
-        this.props.history.push(`/iniciar-sesion`);
-      });
-    } else {
-      swal(
-        "¡Atención!",
-        "La contraseña actual ingresada es incorrecta.",
-        "warning"
-      );
+    const res = await put_request(`updatePassword`, passwords);
+    if (res.status) {
+      const msg = res.data.msg;
+      if (msg !== "Error") {
+        swal(
+          "¡Listo!",
+          "Contraseña actualizada exitosamente. Por motivos de seguridad, se cerrará su sesión",
+          "success"
+        ).then(() => {
+          localStorage.clear();
+          this.props.history.push(`/iniciar-sesion`);
+        });
+      } else {
+        swal(
+          "¡Atención!",
+          "La contraseña actual ingresada es incorrecta.",
+          "warning"
+        );
+      }
     }
   }
 }
